tune intake: Reduce speed of intake when scoring in low goal
We noticed that when scoring low goal the blocks would go too fast, rolling out the other end of the low goal. To fix this, we reduced the speed of the first stage.

Because the first stage is slower, we also have to slow down the second and third stage to avoid clashing of blocks.